+++ src/main/java/duke/exception/InvalidDateTimeException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class InvalidDateTimeException extends DukeException {
+    public InvalidDateTimeException() {
+        super("Please enter a date of the format yyyy-mm-dd!");
+    }
+}
