+++ src/main/java/duke/exception/InvalidTaskDeletionException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class InvalidTaskDeletionException extends DukeException {
+    public InvalidTaskDeletionException() {
+        super("OOPS! You are trying to delete a non-existent task!");
+    }
+}
