+++ src/main/java/duke/exception/NoCommandException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class NoCommandException extends DukeException {
+    public NoCommandException() {
+        super("OOPS! I'm sorry, but I don't know what that means :(");
+    }
+}
