+++ src/main/java/duke/exception/NoDescriptionAndTimeException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class NoDescriptionAndTimeException extends DukeException {
+    public NoDescriptionAndTimeException(String string){
+        super("OOPS! The description & time of " + string +" cannot\nbe empty.");
+    }
+}
