+++ src/main/java/duke/exception/NoTimeException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class NoTimeException extends DukeException {
+    public NoTimeException(String string){
+        super("OOPS! The time of " + string +" cannot be empty.");
+    }
+}
