+++ src/main/java/duke/task/Deadline.java
@@ -1,11 +1,10 @@
-import java.time.LocalDateTime;
-import java.util.Scanner;
+package duke.task;
 
 public class Deadline extends Task {
 
     protected String by;
 
-    public Deadline(String description, String by) throws DukeException {
+    public Deadline(String description, String by) {
         super(description, "deadline");
         this.by = by;
     }
