+++ src/main/java/duke/task/Event.java
@@ -1,9 +1,9 @@
-import java.util.Scanner;
+package duke.task;
 
-public class Event extends Task{
+public class Event extends Task {
     protected String at;
 
-    public Event(String description, String at) throws DukeException {
+    public Event(String description, String at) {
         super(description, "event");
         this.at = at;
     }
