+++ text-ui-test/runtest.bat
@@ -7,7 +7,7 @@ REM delete output from previous run
 if exist ACTUAL.TXT del ACTUAL.TXT
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java  -Xlint:none -d ..\bin ..\src\main\java\*.java
+javac  -cp ..\src\main\java  -Xlint:none -d ..\bin ..\src\main\java\duke\exception\*.java ..\src\main\java\duke\task\*.java ..\src\main\java\duke\util\*.java ..\src\main\java\duke\main\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
